(chore) move tests in hiring_staff feature To ensure any other changes are transferred over and delete redundant files
