--- app/views/hiring_staff/vacancies/application_details.html.haml
@@ -1,26 +1,0 @@
-%h1.heading-large
-  = t('vacancies.publish_heading', school: @school.name)
-  %span.heading-secondary
-    Step 3 of 3
-= render 'error_messages', errors: @vacancy.errors
-= simple_form_for [@school, @vacancy], url: school_vacancy_path(@school.id, @vacancy), html: { class: 'vacancy-form' } do |f|
-  = hidden_field_tag :from, :application_details
-  = hidden_field_tag :next, :review
-  %h2.heading-medium
-    = t('vacancies.application_details')
-
-  .grid-row
-    .column-one-half
-      = f.input :contact_email,
-                label: t('vacancies.contact_email'),
-                hint: t('vacancies.form_hints.contact_email'),
-                as: :email,
-                required: true
-
-      %div.form-group
-        = f.gov_uk_date_field :publish_on, legend_text: t('vacancies.publication_date'), legend_class: 'form-label-bold', form_hint_text: t('vacancies.form_hints.publication_date')
-
-      %div.form-group
-        = f.gov_uk_date_field :expires_on, legend_text: t('vacancies.deadline_date'), legend_class: 'form-label-bold', form_hint_text: t('vacancies.form_hints.deadline_date')
-
-      = f.button :submit, t('buttons.save_and_continue')
