--- app/views/hiring_staff/vacancies/candidate_specification.html.haml
@@ -1,40 +1,0 @@
-%h1.heading-large
-  = t('vacancies.publish_heading', school: @school.name)
-  %span.heading-secondary
-    Step 2 of 3
-= render 'error_messages', errors: @vacancy.errors
-= simple_form_for [@school, @vacancy], url: school_vacancy_path(@school.id, @vacancy), html: { class: 'vacancy-form' } do |f|
-  = hidden_field_tag :from, :candidate_specification
-  = hidden_field_tag :next, (params[:next] || :application_details)
-  %h2.heading-medium
-    = t('vacancies.candidate_specification')
-
-  .grid-row
-    .column-one-half
-      = f.input :essential_requirements,
-                label: t('vacancies.essential_requirements'),
-                as: :text,
-                hint: t('vacancies.form_hints.essential_requirements'),
-                input_html: {rows: 10},
-                required: true
-      = f.input :education,
-                label: t('vacancies.education'),
-                as: :text,
-                hint: t('vacancies.form_hints.education'),
-                input_html: {rows: 5},
-                required: false
-      = f.input :qualifications,
-                label: t('vacancies.qualifications'),
-                as: :text,
-                hint: t('vacancies.form_hints.qualifications'),
-                input_html: {rows: 5},
-                required: false
-      = f.input :experience,
-                label: t('vacancies.experience'),
-                as: :text,
-                hint: t('vacancies.form_hints.experience'),
-                input_html: {rows: 5},
-                required: false
-
-      = f.input :school_id, as: :hidden
-      = f.button :submit, t('buttons.save_and_continue')
