+++ config/routes.rb
@@ -6,18 +6,10 @@ Rails.application.routes.draw do
   resources :vacancies, only: %i[index show]
   resource :sessions, controller: 'hiring_staff/sessions'
   resources :schools, only: %i[index show edit update], controller: 'hiring_staff/schools' do
-    resources :vacancies, only: %i[index new create update edit destroy show], controller: 'hiring_staff/vacancies' do
-      # Legacy form routing copied over
-      get 'review', on: :member
-      put 'publish', on: :member
-      get 'published', on: :member
-      resource :job_specification, only: :show, controller: 'hiring_staff/vacancies', action: :job_specification
-      resource :candidate_specification,
-               only: :show, controller: 'hiring_staff/vacancies',
-               action: :candidate_specification
-      resource :application_details, only: :show, controller: 'hiring_staff/vacancies', action: :application_details
-    end
-    resource :vacancies, only: %i[new create], controller: 'hiring_staff/vacancies' do
+    get 'search', on: :collection
+    resources :vacancies, only: %i[new index destroy delete show], controller: 'hiring_staff/vacancies' do
+      get 'review'
+      post :publish, to: 'hiring_staff/vacancies/publish#create'
     end
 
     resource :vacancy, only: [] do
@@ -27,14 +19,6 @@ Rails.application.routes.draw do
       post :candidate_specification, to: 'hiring_staff/vacancies/candidate_specification#create'
       get :application_details, to: 'hiring_staff/vacancies/application_details#new'
       post :application_details, to: 'hiring_staff/vacancies/application_details#create'
-    end
-
-    resources :vacancy, only: [:show], controller: 'schools/vacancies' do
-      get 'review', to: 'hiring_staff/vacancies#review'
-      post :publish, to: 'hiring_staff/vacancies/publish#create'
-    end
-
-    resources :vacancies, only: [:new], controller: 'schools/vacancies' do
     end
   end
