+++ spec/support/vacancy_helpers.rb
@@ -1,29 +1,4 @@
 module VacancyHelpers
-  def fill_in_job_spec_fields(vacancy)
-    fill_in 'vacancy[job_title]', with: vacancy.job_title
-    fill_in 'vacancy[headline]', with: vacancy.headline
-    fill_in 'vacancy[job_description]', with: vacancy.job_description
-    select vacancy.working_pattern.humanize, from: 'vacancy[working_pattern]'
-    fill_in 'vacancy[minimum_salary]', with: vacancy.minimum_salary
-    fill_in 'vacancy[maximum_salary]', with: vacancy.maximum_salary
-    click_button 'Save and continue'
-  end
-
-  def fill_in_candidate_specification_fields(vacancy)
-    fill_in 'vacancy[essential_requirements]', with: vacancy.essential_requirements
-  end
-
-  def fill_in_application_details_fields(vacancy)
-    fill_in 'vacancy[expires_on_dd]', with: Faker::Business.credit_card_expiry_date.day
-    fill_in 'vacancy[expires_on_mm]', with: Faker::Business.credit_card_expiry_date.strftime('%m')
-    fill_in 'vacancy[expires_on_yyyy]', with: Faker::Business.credit_card_expiry_date.year
-    fill_in 'vacancy[publish_on_dd]', with: Time.zone.today.day
-    fill_in 'vacancy[publish_on_mm]', with: Time.zone.today.strftime('%m')
-    fill_in 'vacancy[publish_on_yyyy]', with: Time.zone.today.year
-    fill_in 'vacancy[contact_email]', with: vacancy.contact_email
-  end
-
-  # rubocop:disable Metrics/AbcSize
   def fill_in_job_specification_form_fields(vacancy)
     fill_in 'job_specification_form[job_title]', with: vacancy.job_title
     fill_in 'job_specification_form[headline]', with: vacancy.headline
@@ -41,7 +16,6 @@ module VacancyHelpers
     fill_in 'job_specification_form[ends_on_mm]', with: vacancy.ends_on.strftime('%m')
     fill_in 'job_specification_form[ends_on_yyyy]', with: vacancy.ends_on.year
   end
-  # rubocop:enable Metrics/AbcSize
 
   def fill_in_candidate_specification_form_fields(vacancy)
     fill_in 'candidate_specification_form[essential_requirements]', with: vacancy.essential_requirements
@@ -50,7 +24,6 @@ module VacancyHelpers
     fill_in 'candidate_specification_form[experience]', with: vacancy.experience
   end
 
-  # rubocop:disable Metrics/AbcSize
   def fill_in_application_details_form_fields(vacancy)
     fill_in 'application_details_form[contact_email]', with: vacancy.contact_email
     fill_in 'application_details_form[expires_on_dd]', with: vacancy.expires_on.day
@@ -60,9 +33,7 @@ module VacancyHelpers
     fill_in 'application_details_form[publish_on_mm]', with: vacancy.publish_on.strftime('%m')
     fill_in 'application_details_form[publish_on_yyyy]', with: vacancy.publish_on.year
   end
-  # rubocop:enable Metrics/AbcSize
 
-  # rubocop:disable Metrics/AbcSize
   def verify_all_vacancy_details(vacancy)
     expect(page).to have_content(vacancy.job_title)
     expect(page).to have_content(vacancy.headline)
@@ -86,7 +57,6 @@ module VacancyHelpers
     expect(page).to have_content(vacancy.expires_on)
     expect(page).to have_content(vacancy.publish_on)
   end
-  # rubocop:enable Metrics/AbcSize
 
   def expect_schema_property_to_match_value(key, value)
     expect(page).to have_selector("meta[itemprop='#{key}'][content='#{value}']")
